Reverted master to c24b503

+++ pingpong/bluetooth.h
@@ -31,7 +31,7 @@ namespace Bluetooth {
    /**
     * We can't use a timer :D /s
     */
-   void listen(unsigned long listenTimeout, unsigned long acceptTimeout = BLUETOOTH_DEFAULT_TIMEOUT);
+   void listen(unsigned int listenTimeout, unsigned int acceptTimeout = BLUETOOTH_DEFAULT_TIMEOUT);
 
 
   /**
